Fix Newark, NJ population numbers (#26)
Newark, NJ, currently has no population/per-capita numbers because of a bug in the code.

It's kind of an interesting bug—Newark happens to have the same place FIPS code as New York City, 51000 (remember that FIPS are only unique within a state!). So the special handling code for NYC boroughs was setting the place_or_county_code for Newark to be null, resulting in a failed join when adding population data to the places permits dataset. The fix is also to also check that the state code = 36 (NY) in the NYC-specific handling code.