Distinguish single-city MSAs from cities in comparison view (#53)
Currently metro names are displayed in the comparison view (both the dropdown and the chart label) as "[MSA name], [state]". This means that metros named after just one city, such as the Madison, WI metro area, are displayed as just "Madison, WI", which means that we can't distinguish between the Madison metro area and Madison the city.
It also messes up the plot, only one line is shown rather than two if you select both the Madison metro area and the city.

This fixes this mistake by, in the comparison view only, appending "MSA" or "CSA" to the metro name in both the dropdown and the chart view.

This also makes a minor change to the "Metros" bar plot change, renaming "CBSA" to "MSA". And it removes the Micropolitan Statistical Areas from the dataset, since most people don't care about them.