Fix off-by-one error in tooltip date
In the "Comparisons" graph with the "No averaging" option, the tooltip
that you get when you mouse over a point was giving me a "binned_year"
value one less than it should.  Turns out Vega-Lite defaults to
displaying all dates in local time.  I'm in UTC-5, and the x values are
all midnight January first in the given year, so subtracting 5 hours are
enough that "%Y" displays the previous year.

There's gotta be a simpler way to do all this, but this works.