Fix duplicates (#66)
Currently we drop the place type (i.e. town, village, etc.) when cleaning up the place names.

If there are multiple place types for the same (place name, state) tuple, we combine them together into one plot. This is very wrong. (I'm not sure how we're getting the populations—maybe choosing an arbitrary one? And before #64 we were dropping all of these "duplicates"!)

This switches to a better solution: if there is only one place type for the (place name, state) tuple, don't include the place type in the name. If there are multiple (e.g. Marcellus town, NY, and Marcellus village, NY), include the full name for each.