Map Montreal arrondisements to Montreal city (#67)
Montréal reports its numbers by borough/arrondisement, since each arrondisement has its own local government/permitting authority. We need to map the arrondisement SGC code to the one for the city of Montréal for it to join successfully to the crosswalk.

This makes the data look semi-reasonable for 2003-2021—earlier Montréal was entirely missing. 2000-2002 is still missing, I need to dig more for that