Make bars wider (#71)
This makes the Canada ones look much less stupid, and also makes the US ones look better.

Arguably I should go all the way to 95% or whatever the default is (if I just remove the `width` parameter). I'm hesitant to change the UI so drastically though, so I'll just leave it like this for now.