Remove duplicate row for 2020 data (#79)
The bar for 2020 shouldn't be twice as high as 2019 and 2021...

This is a bug introduced in #78. We have two population rows for 2020 (one from the 2010s dataset, one from the 2020s dataset), and when it gets joined to the BPS data we get two identical rows for 2020 for each state. This makes the bar look twice as high