Make "download data" contents not "[]"/"undefined" (#89)
Currently the download link is randomly broken because when the `url.current` ref changes the `<a>` doesn't necessarily re-render. Using `setState` ensures that the `<a>` re-renders immediately when the data changes.

This currently causes the behavior where the download link gives you a file containing the initial state of `data` (`[]` on the comparisons page, `undefined` on the bar plots) rather than the most up-to-date contents of `data` (which should be populated once data is successfully fetched).